feat: commit updates to host npm package manifest
Previously this was disabled because this action overwrote the
host (root) package.json file with its own. If let semantic-release
commit the package.json, we'd be overwriting the host's package.json
with this action's!

Now that this action uses a temporary directory to contain its
npm package manifest, we can re-enable committing the host
package.json and package-lock.json without fear of over-writing.